fix most read page tests
